rename: theme.ts를 index.ts로 변경
