+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from 'next/app';
 import { RecoilRoot } from 'recoil';
 import { Layout } from 'components/template';
 import { ThemeProvider } from '@emotion/react';
-import theme from 'styles/theme/theme';
+import theme from 'styles/theme';
 
 function App({ Component, pageProps }: AppProps) {
   return (
